+++ node-require/multiply.js
@@ -1,6 +1,6 @@
 exports.times = function multiplyValues(a, b) {
-  if (isNaN(parseInt(a)) || isNaN(parseInt(b))) {
+  if (isNaN(parseFloat(a)) || isNaN(parseFloat(b))) {
     return 'invalid input';
   }
-  return parseInt(a) * parseInt(b);
+  return parseFloat(a) * parseFloat(b);
 };
