+++ node-require/subtract.js
@@ -1,8 +1,8 @@
 exports.subtract = function subtractValues(a, b) {
 
-  if (isNaN(parseInt(a)) || isNaN(parseInt(b))) {
+  if (isNaN(parseFloat(a)) || isNaN(parseFloat(b))) {
     return 'invalid input';
   }
-  return parseInt(a) - parseInt(b);
+  return parseFloat(a) - parseFloat(b);
 
 };
